LayoutTests/canvas/philip/tests/2d.line.join.invalid-expected.txt

< [index] >
2d.line.join.invalid
Setting lineJoin to invalid values is ignored
References: 2d.lineJoin.invalid Bugzilla
Actual output:
Passed
